hackathon-todo-2.0-phase3/pyproject.toml
[project]
name = "hackathon-todo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "sqlmodel",
    "uvicorn",
    "alembic",
    "pydantic-settings",
    "python-jose[cryptography]",
    "python-multipart",
    "psycopg2-binary",
]
